Moving edit mode tests of data tracks
