+++ Tests/EditMode/DataTrackTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 054a3a702298f40128bef9d317aa70a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
